Finished re-implementation of the internal booking API for service professional bookings and shared code with other internal updates to bookings and calculations for client bookings. Added enumeration for CalendarAvailabilityType to reduce the 'magic numbers' on the code.

+++ web/App_Code/LcEnum/BookingType.cs
@@ -8,7 +8,7 @@ namespace LcEnum
     /// <summary>
     /// 
     /// </summary>
-    public enum BookingType : short
+    public enum BookingType : int
     {
         marketplaceBooking = 1,
         bookNowBooking = 2,
